Move prepare_ner_data to the ner script directory
